Change the serializer depend of action method.
With these change the method call "get_serializer_class" for send different serializer depend in what method is call.

It isn't the best practice because when you has an entity and it has relation with other entity you need pass the PK and search the detail of other entity on other resource/endpoint.